+++ benches/gossips.rs
@@ -52,7 +52,6 @@ fn gossip_poi_bench(c: &mut Criterion) {
         server_port: None,
         log_format: String::from("pretty"),
         persistence_file_path: None,
-        log_format: String::from("pretty"),
     });
     _ = black_box(CONFIG.set(Arc::new(SyncMutex::new(config))));
